fix(models): satisfy backend lint ratchets
